calculations (line drawing work in progress)

+++ src/casting_utils.c
@@ -6,7 +6,7 @@
 /*   By: dmalacov <dmalacov@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/02/23 11:04:01 by dmalacov      #+#    #+#                 */
-/*   Updated: 2023/03/02 17:07:14 by dmalacov      ########   odam.nl         */
+/*   Updated: 2023/03/07 17:40:24 by dmalacov      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -75,7 +75,7 @@ int	facing_what(t_point *step, int axis)
 		return (NORTH);
 }
 
-int	is_wall(t_data *data, int x, int y)
+int	is_wall(t_data *data, double x, double y)
 {
 	const char	map[10][31] = \
 	{"111111111111111111111111111111", \
@@ -90,7 +90,7 @@ int	is_wall(t_data *data, int x, int y)
 	"111111111111111111111111111111"};
 	
 	// printf("* is wall: map is %c\n", map[(int)round(data->goat->y) + y][(int)round(data->goat->x) + x]);
-	if (map[(int)floor(data->goat->y) + y][(int)floor(data->goat->x) + x] \
+	if (map[(int)floor(data->goat->y + y)][(int)floor(data->goat->x + x)] \
 	== '1')
 		return (1);
 	else
